@@ -7,7 +7,7 @@ import '../impl.dart';
 class DirUtils implements LocalStorageImpl {
   DirUtils(this.fileName, [this.path]);
 
-  final String path, fileName;
+  final String? path, fileName;
 
   StreamController<Map<String, dynamic>> storage =
       StreamController<Map<String, dynamic>>();
@@ -38,7 +38,7 @@ class DirUtils implements LocalStorageImpl {
   }
 
   @override
-  Future<void> init([Map<String, dynamic> initialData]) {
+  Future<void> init([Map<String, dynamic>? initialData]) {
     throw PlatformNotSupportedError();
   }
 
